+++ rosariosis_mysql.sql
@@ -1659,7 +1659,7 @@ CREATE TABLE school_fields (
 CREATE TABLE school_gradelevels (
     id integer NOT NULL AUTO_INCREMENT PRIMARY KEY,
     school_id integer NOT NULL,
-    short_name varchar(3),
+    short_name varchar(4),
     title varchar(50) NOT NULL,
     next_grade_id integer,
     sort_order numeric,